Move StepPiece and LinearPiece classes to base pieces module
