+++ src/pieces/base.js
@@ -16,9 +16,10 @@ limitations under the License.
 
 
 const { Relation, PieceSquares } = require('../relations');
+const { SquareCoordinates } = require('../square');
 
 
-/** Base chess piece class. */
+/** Base piece class. */
 class Piece {
 
     static WHITE = 'white';
@@ -269,6 +270,77 @@ class Piece {
 }
 
 
+/** Base step piece class. */
+class StepPiece extends Piece {
+
+    /**
+     * Creation.
+     * @param {string} color - Either white or black.
+     * @param {Square} square - Square piece place to.
+     * @param {string} kind - One of `Piece.ALL_KINDS`.
+     * @param {boolean} [refresh=true] - Whether refresh board or not.
+     */
+    constructor(color, square, kind, refresh=true) {
+        super(color, square, kind, refresh);
+    }
+
+    /**
+     * Get step piece squares by piece action.
+     * @param {Object[]} stepPoints - Squares points.
+     * @param {integer} stepPoints.x - X square coordinate.
+     * @param {integer} stepPoints.y - Y square coordinate.
+     */
+    _getStepSquares(stepPoints) {
+        this._refreshSquareFinder();
+        for (let stepPoint of stepPoints) {
+            let x = this.square.coordinates.x + stepPoint.x;
+            let y = this.square.coordinates.y + stepPoint.y;
+            if (!SquareCoordinates.correctCoordinates(x, y)) continue;
+
+            this._nextSquareAction(this.board.squares.getFromCoordinates(x, y));
+        }
+    }
+}
+
+
+/** Base linear piece class. */
+class LinearPiece extends Piece {
+
+    /**
+     * Creation.
+     * @param {string} color - Either white or black.
+     * @param {Square} square - Square piece place to.
+     * @param {string} kind - One of `Piece.ALL_KINDS`.
+     * @param {boolean} [refresh=true] - Whether refresh board or not.
+     */
+    constructor(color, square, kind, refresh=true) {
+        super(color, square, kind, refresh);
+    }
+
+    /**
+     * Get step piece squares by piece action.
+     * @param {Object[]} directions - Squares directions.
+     * @param {integer} directions.x - X direction delta.
+     * @param {integer} directions.y - Y direction delta.
+     */
+    _getLinearSquares(directions) {
+        for (let direction of directions) {
+            this._refreshSquareFinder();
+            let x = this.square.coordinates.x + direction.x;
+            let y = this.square.coordinates.y + direction.y;
+            while (SquareCoordinates.correctCoordinates(x, y)) {
+                this._nextSquareAction(this.board.squares.getFromCoordinates(x, y));
+                if (this.endOfALine) break;
+                x += direction.x;
+                y += direction.y;
+            }
+        }
+    }
+}
+
+
 module.exports = {
     Piece: Piece,
+    StepPiece: StepPiece,
+    LinearPiece: LinearPiece,
 };
